Gave TreeFeller a cooldown
