--- progression/src/main/java/me/mykindos/betterpvp/progression/profession/skill/woodcutting/TreeFeller.java
@@ -1,38 +1,0 @@
-package me.mykindos.betterpvp.progression.profession.skill.woodcutting;
-
-import com.google.inject.Inject;
-import com.google.inject.Singleton;
-import me.mykindos.betterpvp.progression.Progression;
-import org.bukkit.Material;
-import org.bukkit.inventory.ItemFlag;
-
-@Singleton
-public class TreeFeller extends WoodcuttingProgressionSkill {
-
-    @Inject
-    public TreeFeller(Progression progression) {
-        super(progression);
-    }
-
-    @Override
-    public String getName() {
-        return "Tree Feller";
-    }
-
-    @Override
-    public String[] getDescription(int level) {
-        return new String[]{
-                "Cut down an entire tree by chopping a single log",
-        };
-    }
-
-    @Override
-    public Material getIcon() {
-        return Material.GOLDEN_AXE;
-    }
-
-    @Override
-    public ItemFlag getFlag() {
-        return ItemFlag.HIDE_ATTRIBUTES;
-    }
-}
